Extract `useNewForwardingAddresses`
